* some fixes

+++ src/restore.cpp
@@ -239,7 +239,7 @@ void save_game( const String filename, GAME_DATA *game_data )
 			SAVE( units.unit[i].weapon[f].delay );
 			SAVE( units.unit[i].weapon[f].time );
 			SAVE( units.unit[i].weapon[f].target_pos );
-			int g = units.unit[i].weapon[f].target ? ( (units.unit[i].weapon[f].state & WEAPON_FLAG_WEAPON) ? ((WEAPON*)units.unit[i].weapon[f].target)->idx : ((UNIT*)units.unit[i].weapon[f].target)->idx ) : -1;
+			int g = units.unit[i].weapon[f].target ? (int)( (units.unit[i].weapon[f].state & WEAPON_FLAG_WEAPON) ? ((WEAPON*)units.unit[i].weapon[f].target)->idx : ((UNIT*)units.unit[i].weapon[f].target)->idx ) : -1;
 			SAVE( g );
 			SAVE( units.unit[i].weapon[f].data );
 			SAVE( units.unit[i].weapon[f].flags );
@@ -306,7 +306,7 @@ void save_game( const String filename, GAME_DATA *game_data )
 				{
 				case MISSION_ATTACK:
 					{
-						int p = cur->p ? ( (cur->flags&MISSION_FLAG_TARGET_WEAPON) ? ((WEAPON*)(cur->p))->idx : ((UNIT*)(cur->p))->idx) : -1;
+						int p = cur->p ? ( (cur->flags&MISSION_FLAG_TARGET_WEAPON) ? (int)((WEAPON*)(cur->p))->idx : ((UNIT*)(cur->p))->idx) : -1;
 						SAVE( p );
 					}
 					break;
@@ -320,7 +320,7 @@ void save_game( const String filename, GAME_DATA *game_data )
 				case MISSION_REVIVE:
 				case MISSION_GET_REPAIRED:
 					{
-						int p = cur->p ? ((UNIT*)(cur->p))->idx : -1;
+						int p = cur->p ? (int)((UNIT*)(cur->p))->idx : -1;
 						SAVE( p );
 					}
 					break;
